Added `SettingWithCopyWarning`
Hi Sean, 

When reviewing my code for Google Trends, I found I made one mistake (`SettingWithCopyWarning`) a couple of times. Because my code still works with these warnings, I did not pay much attention to it initially. But after digging into it, I found it is a very serious error that may change the results unexpectedly. That is why I added it to the guide. But the way I illustrate it still needs to be improved for sure...